requirements.txt
openai
pandas
python-dotenv
streamlit
openyxl
